pyproject.toml
[project]
name = "s3-python-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.0.0",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]
